fix: deadlock in the logic
Fetching the stale branches in unbuffered channel and hence the buffer
size is not defined it is not blocked on the receiver to pop from the
channel. Hence somethings racefull condition is happening because data
is not present in the correct fashion. Using a buffered channel makes it
much more faster and simpler to handle the concurrency.