fix: remove blocking of channel
We were comsuming the output from channel after the closing of channel
which is creating a sort of stuck program if the distribution in array
is not to a particular order.

Moved the implementation to use a go-routine so that it cannot be
blocked for a large amount of time.